+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This changelog was created using the `clu` binary
 (https://github.com/MalteHerrmann/changelog-utils).
 -->
+
 # Changelog
 
 ## Unreleased
@@ -30,3 +31,4 @@ This changelog was created using the `clu` binary
 - (proto) [#14](https://github.com/evmos/os/pull/14) Add Protobufs and adjust scripts.
 - (eip-712) [#13](https://github.com/evmos/os/pull/13) Add EIP-712 package.
 - (ci) [#12](https://github.com/evmos/os/pull/12) Add CI workflows, configurations, Makefile, License, etc.
+- (tests) [#65](https://github.com/evmos/os/pull/65) Add tests for bank wrapper in EVM module with mock.